code/p5/requirements.txt
# P5 — Change Detection + UNESCO + Visualization
# Pomzadoya hackathon, Modül A v2

folium>=0.15.0
leafmap>=0.31.0
shapely>=2.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
rasterio>=1.3.8
streamlit>=1.30.0
geopandas>=0.14.0
requests>=2.31.0
branca>=0.7.0
numpy>=1.24.0
pandas>=2.0.0

# transitive but pinned for reproducibility
fiona>=1.9.0
pyproj>=3.6.0
matplotlib>=3.7.0
pillow>=10.0.0

# P3 inference fn'ini import etmek için (T5.10 historical RAW)
torch>=2.0.0
torchvision>=0.15.0
